pyproject.toml
[tool.poetry]
name = "vatglobaldevopsinterview"
version = "0.1.0"
description = ""
authors = ["Christo Goosen <christogoosen@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.5"
django-rest = "^0.8.6"
django-q = "^1.3.4"
django-cors-headers = "^3.7.0"
djangorestframework = "^3.12.2"
markdown = "^3.3.3"
django-filter = "^2.4.0"
psutil = "^5.8.0"
drf-yasg = "^1.20.0"
djangorestframework-camel-case = "^1.2.0"
names = "^0.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
